requirements.txt
selenium==4.5.0
boto3==1.26.2
python-dotenv